move models to db package
