+++ daemon/database/models/StatusEnum.go
@@ -1,4 +1,4 @@
-package swap
+package models
 
 import (
 	"database/sql/driver"
@@ -13,17 +13,14 @@ const (
 	StatusFailed    SwapStatus = "failed"
 )
 
-// IsValid checks if the SwapStatus value is valid
 func (s SwapStatus) IsValid() bool {
 	return s == StatusPending || s == StatusCompleted || s == StatusFailed
 }
 
-// String returns the string representation
 func (s SwapStatus) String() string {
 	return string(s)
 }
 
-// Scan implements the sql.Scanner interface for SwapStatus
 func (s *SwapStatus) Scan(value interface{}) error {
 	str, ok := value.(string)
 	if !ok {
@@ -34,7 +31,6 @@ func (s *SwapStatus) Scan(value interface{}) error {
 	return nil
 }
 
-// Value implements the driver.Valuer interface for SwapStatus
 func (s SwapStatus) Value() (driver.Value, error) {
 	return string(s), nil
 }
